feat(Menu): Introduce MenuItemLink component
