+++ src/components/Menu/index.ts
@@ -1,2 +1,3 @@
 export { default } from './Menu'
 export { default as MenuItem } from './MenuItem'
+export { default as MenuItemLink } from './MenuItemLink'
