+++ src/index.ts
@@ -24,6 +24,7 @@ export { default as List } from './components/List'
 export { ListItem } from './components/List'
 export { default as Menu } from './components/Menu'
 export { MenuItem } from './components/Menu'
+export { MenuItemLink } from './components/Menu'
 export { default as Provider } from './components/Provider'
 export { default as ProviderConsumer } from './components/Provider/ProviderConsumer'
 export { default as RadioGroup } from './components/RadioGroup'
