+++ test/specs/components/Menu/MenuItemLink-test.tsx
@@ -1,0 +1,8 @@
+// import * as React from 'react'
+
+import MenuItemLink from 'src/components/Menu/MenuItemLink'
+import { isConformant } from 'test/specs/commonTests'
+
+describe('MenuItemLink', () => {
+  isConformant(MenuItemLink)
+})
